4.Authentication&Authorization/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic
python-multipart
